python_/requirements.txt
appier==1.17.8
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
google-api==0.1.12
googleapis-common-protos==1.5.3
grpcio==1.15.0
grpcio-tools==1.15.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.6.1
pylint==1.9.3
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
